+++ services/storage/src/simcore_service_storage/api/rpc/_async_jobs.py
@@ -6,7 +6,7 @@ from models_library.api_schemas_rpc_data_export.async_jobs import (
     AsyncJobRpcResult,
     AsyncJobRpcStatus,
 )
-from models_library.api_schemas_storage.data_export_tasks import (
+from models_library.api_schemas_storage.data_export_async_jobs import (
     DataExportTaskAbortOutput,
 )
 from servicelib.rabbitmq import RPCRouter
