+++ services/storage/src/simcore_service_storage/api/rpc/_data_export.py
@@ -2,7 +2,7 @@ from uuid import uuid4
 
 from fastapi import FastAPI
 from models_library.api_schemas_rpc_data_export.async_jobs import AsyncJobRpcGet
-from models_library.api_schemas_storage.data_export_tasks import (
+from models_library.api_schemas_storage.data_export_async_jobs import (
     DataExportTaskStartInput,
 )
 from servicelib.rabbitmq import RPCRouter
